age change

See merge request capstone-project-PlantSketch/plantsketch!92

+++ src/main/java/plantsketch/Plant.java
@@ -92,7 +92,7 @@ public class Plant {
     // Plant setters
     // ============================
 
-    public void setAge(int currentAge) {
+    public void setAge(float currentAge) {
         this.currentAge = currentAge;
     }
